requirements.txt
requests>=2.31.0
pyyaml>=6.0
click>=8.1.0
python-dateutil>=2.8.0
jinja2>=3.1.0